msm: devfreq_cpubw: Add devfreq driver for CPU to DDR bandwidth voting
This driver registers itself as a devfreq device that allows devfreq
governors to make CPU to DDR IB/AB bandwidth votes. This driver allows the
governors to be agnostic of the bandwidth voting APIs, the number of CPU
master ports or DDR slave ports, the actual port numbers, the system
topology, the available DDR frequencies, etc.

Change-Id: If055ddd580afd41f9668b111e6c09a047488b2e0
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>
Signed-off-by: Domenico Marino <johnlemon.lol@gmail.com>

Conflicts:
	arch/arm/mach-msm/Kconfig